Adjusted the queries based on proper database implementation
Was initially unaware of the structure of the Users table in our database so I had created a separate UserProfile table to handle the image and bio even though the users table had these columns already.
With this in mind I adjusted the register user api to insert all relavent profile data into the Users table, including a default image and bio.

I modified the UserProfile component to deal with conditional rendering to display a loading icon in the case of data retrieval happening after the component mounts. Also added a default state of profileData to include the desired properties to prevent possible undefined access errors.