Added task editing feature
Firat I created an EditTask component that displayed a form that was pre-set to have the values of the current task. In order to display this component I created a button on each task in the TaskList component that when clicked, used the state setCurrentTask(task) that was managed in the GroupPage
Each field in the form was managed with state and upon change of that data the state was set to the target value. Upon submission the edit function is called which waits for a response from editTask()

editTask(channel, task, formData) was defined in the actions.js file to defined a PUT request to update the existing data for the task specified. The channel and task are necessary parameters in order to perform an efficient SQL query on the database. Within this function it performs a fetch to a dynamic API endpoint edit-task/[taskID] which I describe below.

The API endpoint edit-task/[taskID] extracts the taskID from the url segment and sets title, description, and dueDate of the task and the join url of the group to the data sent from the PUT request. The use, task, and channel are verified via an SQL query and the response is sent back.