Added api endpoint functionality for task editing
First I added the implementation for the edit function which is called on upon submission of the editing form. Within this function I added channel and id parameters to deal with SQL queries in the endpoint. This function awaits a response from the editTask function defined in the actions file

In the editTask unction I defined a data object to get the title, description, and due_date from the form. I then defined query paramters to hold the channel and id params to fetch from the edit-task endpoint. The request method was a PUT in order to update the tasks.

Within the edit-task endpoint I extracted te url and task id from the query parameters, and also extracted the title,description, and due_date. I then performed an update query on the task by setting the proper columns and checking it belonged to the correct group and user.